coding phase

+++ src/main/java/drinkshop/domain/Stoc.java
@@ -7,7 +7,7 @@ public class Stoc {
     private double cantitate;
     private double stocMinim;
 
-    public Stoc(int id, String ingredient, int cantitate, int stocMinim) {
+    public Stoc(int id, String ingredient, double cantitate, double stocMinim) {
         this.id = id;
         this.ingredient = ingredient;
         this.cantitate = cantitate;
